move merkle tree to data_model
